move examples into strories folder
