--- src/hooks.stories.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { InputPreview } from "./InputPreview";
-
-export default {
-  title: "Input preview",
-  component: InputPreview,
-};
-
-export const Default = (): React.ReactNode => <InputPreview />;
